refactor(music): music provider now handles getStream and seeking
The provider should now return it's own instance to the linkable song, allowing the song to access its provider data / methods.

Seek was also added as a music provider feature, since different providers might seek in different ways. The provider can simply set the song.options.seek if the provider desires to let the connection handle it. If the provider throws an error, the song is not seeked and a message is sentcontaining the error.