fix: disconnect event fired twice
this caused weird issues such as "Last Played Song" button not reappearing because the set history button function was called twice, clearing the previous and new history widgets.

this is especially here because I actually fixed the issues in the library (discord music player). it now correctly emits the disconnect event, where I had to emit it myself previously.