fix: use shardReady events to gracefully handle connection issues
if the power went out, the connection went out too, which led to weird cases where the activity wouldn't be set for the bot user.

more drastic however is that if the bot was added to a server while the bot was active but not online, the guild would not be added to the database and many issue would arrise.